+++ lib/src/features/configuration/presentation/screens/settings_screen.dart
@@ -15,11 +15,11 @@ class SettingsScreen extends ConsumerWidget {
   Widget build(BuildContext context, WidgetRef ref) {
     final languageCode = ref.watch(languageProvider);
     final titles = ref.watch(titlesProvider!);
-    final localeTitles = titles.getTitles(languageCode);
+    final localeTitle = titles.getTitle(languageCode);
 
     return Scaffold(
       appBar: TopBar(
-        localeTitles,
+        localeTitle,
         hasBackButton: true,
       ),
       body: SingleChildScrollView(