+++ lib/src/features/not_found/not_found_screen.dart
@@ -18,10 +18,10 @@ class NotFoundScreen extends ConsumerWidget {
   Widget build(BuildContext context, WidgetRef ref) {
     final languageCode = ref.watch(languageProvider);
     final titles = ref.watch(titlesProvider!);
-    final localeTitles = titles.getTitles(languageCode);
+    final localeTitle = titles.getTitle(languageCode);
 
     return Scaffold(
-      appBar: TopBar(localeTitles),
+      appBar: TopBar(localeTitle),
       body: EmptyPlaceholderWidget(
         message: '404 - Page not found!',
       ),