+++ lib/src/features/questions/presentation/screens/question_prayer_screen.dart
@@ -1,5 +1,6 @@
 import 'package:catechism/src/common_widgets/top_bar.dart';
 import 'package:catechism/src/features/configuration/data/language_provider.dart';
+import 'package:catechism/src/features/configuration/domain/title_data.dart';
 import 'package:catechism/src/features/questions/data/question_provider.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
@@ -18,7 +19,10 @@ class QuestionPrayerScreen extends StatelessWidget {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: TopBar(
-        [],
+        TitleData(
+          title: Text(''),
+          height: 80.0,
+        ),
         hasBackButton: false,
         hasCloseButton: true,
       ),